fix: effects dependencies aren't detected properly in some cases
If the effect has conditions inside it, and some of the branches don't run on the first effect run, some signals don't add the effect to their subscribers list. This fix ensures that each time the effect runs, it adds itself as the `currentEffect`, so the accessed signals will add it to their subscribers list for future runs.